Remove reference repository and document dummy data
