+++ lib/main.dart
@@ -1,22 +1,25 @@
+// lib/main.dart
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'data/inspection_repository.dart';
-import 'data/reference_repository.dart';
 import 'providers/inspection_provider.dart';
 import 'router/app_router.dart';
 
+/// 주요 기능:
+/// - 앱 실행 전 실사 및 참조 더미 데이터를 초기화합니다.
+/// - 전역 [InspectionProvider]를 주입하고 라우터를 구성합니다.
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   final inspectionRepository = InspectionRepository();
-  final referenceRepository = ReferenceDataRepository();
-  final inspectionProvider = InspectionProvider(inspectionRepository, referenceRepository);
+  final inspectionProvider = InspectionProvider(inspectionRepository);
   await inspectionProvider.initialize();
   final router = AppRouter(inspectionProvider).router;
 
   runApp(MyApp(inspectionProvider: inspectionProvider, router: router));
 }
 
+/// Material 3 테마와 GoRouter를 사용하는 최상위 위젯.
 class MyApp extends StatelessWidget {
   const MyApp({super.key, required this.inspectionProvider, required this.router});
